Rename test/util/file.js to trigger.js
